Fix merge conflicts

+++ pkg/minikube/cruntime/docker.go
@@ -89,15 +89,44 @@ func (r *Docker) ListContainers(filter string) ([]string, error) {
 	if err != nil {
 		return nil, err
 	}
-	return strings.Split(content, "\n"), nil
+	var ids []string
+	for _, line := range strings.Split(content, "\n") {
+		if line != "" {
+			ids = append(ids, line)
+		}
+	}
+	return ids, nil
 }
 
-// KillContainers forcibly removes a running pod based on ID
+// KillContainers forcibly removes a running container based on ID
 func (r *Docker) KillContainers(ids []string) error {
+	if len(ids) == 0 {
+		return nil
+	}
+	glog.Infof("Killing containers: %s", ids)
 	return r.Runner.Run(fmt.Sprintf("docker rm -f %s", strings.Join(ids, " ")))
 }
 
-// StopContainers stops a running pod based on ID
+// StopContainers stops a running container based on ID
 func (r *Docker) StopContainers(ids []string) error {
+	if len(ids) == 0 {
+		return nil
+	}
+	glog.Infof("Stopping containers: %s", ids)
 	return r.Runner.Run(fmt.Sprintf("docker stop %s", strings.Join(ids, " ")))
+}
+
+// ContainerLogCmd returns the command to retrieve the log for a container based on ID
+func (r *Docker) ContainerLogCmd(id string, len int, follow bool) string {
+	var cmd strings.Builder
+	cmd.WriteString("docker logs ")
+	if len > 0 {
+		cmd.WriteString(fmt.Sprintf("--tail %d ", len))
+	}
+	if follow {
+		cmd.WriteString("--follow ")
+	}
+
+	cmd.WriteString(id)
+	return cmd.String()
 }
